backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain Ecosystem (CORE REQUIREMENTS)
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langgraph==0.0.20
langserve[all]==0.0.30
langsmith==0.0.69

# AI/ML Libraries
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.36.0
numpy==1.24.3
scikit-learn==1.3.0

# Database & Vector Store
psycopg2-binary==2.9.7
pinecone-client==2.2.4
sqlalchemy==2.0.23
alembic==1.12.1

# Document Processing
pypdf==3.17.0
pdfminer.six==20221105
python-multipart==0.0.6
aiofiles==23.2.1

# HTTP & API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.0